+++ exp1/fitness.cpp
@@ -1,12 +1,12 @@
 #include"fitness.h"
-NNFitness::NNFitness( vector< vector< float > >& input, vector< vector< float > >& target ){
+NNFitness::NNFitness( vector< vector< double > >& input, vector< vector< double > >& target ){
     this->target = transpose( target );
     this->input = transpose( input );
 }
 
 void NNFitness::evaluate( NeuralNetwork<>& nn ){
     size_t NEURONS = nn.getNeurons().size();
-    vector< vector< float > > values;
+    vector< vector< double > > values;
     for( size_t e = 0; e < input.size(); e++ ){
         for( size_t i = 0; i < NEURONS / 2; i++ ){
             nn.setValue( i, input[e][i] );
@@ -18,11 +18,11 @@ void NNFitness::evaluate( NeuralNetwork<>& nn ){
     nn.setFitness( -1 * MSE( values ) );
 }
 
-void NNFitness::setTarget( vector< vector< float > >& target ){
+void NNFitness::setTarget( vector< vector< double > >& target ){
     this->target = transpose( target );
 }
 
-double NNFitness::MSE( vector< vector< float > >& pred ){
+double NNFitness::MSE( vector< vector< double > >& pred ){
     double ret = 0;
     for( size_t i = 0; i < pred.size(); i++ ){
         for( size_t j = 0; j < pred[i].size(); j++ ){
@@ -33,7 +33,7 @@ double NNFitness::MSE( vector< vector< float > >& pred ){
     return ret;
 }
 
-double NNFitness::diff( vector< vector< float > >& pred ){
+double NNFitness::diff( vector< vector< double > >& pred ){
     double ret = 0;
     for( size_t i = 0; i < pred.size(); i++ ){
         for( size_t j = 0; j < pred[i].size(); j++ ){
