+++ exp1/main.cpp
@@ -12,12 +12,12 @@ using std::ofstream;
 #include"fitness.h"
 #include"evolve.h"
 size_t NEURONS = 2050;
-void runRead( vector< vector< float > >&& real, NeuralNetwork<>& nn );
+void runRead( vector< vector< double > >&& real, NeuralNetwork<>& nn );
 
 int main(){
 
-    vector< vector< float > > real = readNumpyArray( "./stft/fund.stft" );
-    vector< vector< float > > target = readNumpyArray( "./stft/Harm1.stft" );
+    vector< vector< double > > real = readNumpyArray( "./stft/fund.stft" );
+    vector< vector< double > > target = readNumpyArray( "./stft/Harm1.stft" );
     NeuralNetwork<> nn( NEURONS, -5, 5, 0.5, 0.5, 0, 0 );
     for( size_t i = 0; i < NEURONS / 2; i++ ){
         nn.addConnection( i, i + NEURONS / 2, 1 );
@@ -35,8 +35,8 @@ int main(){
 }
 
 
-void runRead( vector< vector< float > >&& real, NeuralNetwork<>& nn ){
-    vector< vector< float > > values;
+void runRead( vector< vector< double > >&& real, NeuralNetwork<>& nn ){
+    vector< vector< double > > values;
     for( size_t e = 0; e < real.size(); e++ ){
         for( size_t i = 0; i < NEURONS / 2; i++ ){
             nn.setValue( i, real[e][i] );
